+++ changelog.md
@@ -1,5 +1,6 @@
 ============================================
 - Added connect all notificationTypes on signup
+- Added TopicsPage component
 
 3.7.0-dev
 ============================================